Create subscribers on demand
Add a new vty option and allow to optionally generate a random msisdn,
as well as setting the default NAM:

subscriber-create-on-demand <3-15> (none|cs|ps|both)
subscriber-create-on-demand no-extension

I thought about using an enum or bitfield to store nam_cs and nam_ps,
but I have used booleans instead, to match how it is stored in the
database. Thanks to Vadim for the random MSISDN patch [1], which was
squashed into this one.

[1] Change-Id: I475c71f9902950fa7498855a616e1ec231fad6ac

Related: OS#2542
Change-Id: I0c9fe93f5c24b5e9fefb513c4d049fb7ebd47ecd