db_hlr.c: add db_subscr_exists_by_msisdn()
Check if a subscriber exists without generating an error log entry if
it does not. This is cheaper than db_subscr_get_by_msisdn(), as it
does not fetch the subscriber entry.

subscriber-create-on-demand will use this function to generate
a random unique MSISDN for new subscribers.

TODO: update the unit test!

Change-Id: Ibfbc408c966197682ba2b12d166ade4bfeb7abc2
Related: OS#2542